Add support for GPU direct nv-peer-memory
This commit adds support for building nv peer memory
driver container from sources.

The actual compilation and installation of the module is done in
entrypoint.sh as it is required to compile agains both Nvidia
GPU driver and Mellanox OFED driver.

This requires the container runtime to mount into the container:

1. OFED driver container root fs in /run/mellanox/drivers
2. NVIDIA driver container root fs in /run/nvidia/drivers

And run the container in privileged mode as it will attempt to
reload nv_peer_mem kernel module.

In addition, A patch is intorduced to fix nvidia symver
untill https://github.com/Mellanox/nv_peer_memory/issues/70 is solved