Add support for nvidia-peermem module builtin in GPU driver
GPU-direct container now supports loading of the nvidia-peermem
module which is a part of GPU driver (465 branch)
By default, nv_peer_mem module from the image will be used.
If "ALLOW_MOD_FROM_GPU_DRIVER" environment variable is "true",
entrypoin.sh will check for nvidia_peermem module presence and will use
it if it exists. If module is not found, nv_peer_mem will be used
as a fallback.

Signed-off-by: Yury Kulazhenkov <ykulazhenkov@nvidia.com>